9.115 Authorization. Using the User ID on the Frontend

+++ backend/routes/posts.js
@@ -40,7 +40,8 @@ router.post(
       // _id: req.body.id,        //questa linea non deve esserci! altrim si forza il DB a dare _id = null al nuovo post
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + '/images/' + req.file.filename
+      imagePath: url + '/images/' + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -56,7 +57,7 @@ router.post(
 
 router.put("/:id", checkAuth, multer({ storage: storage }).single("image"), (req, res, next) => {
   let imagePath = req.body.imagePath;
-  if(req.file){
+  if (req.file) {
     const url = req.protocol + '://' + req.get('host');
     imagePath = url + '/images/' + req.file.filename;
   }
@@ -64,12 +65,16 @@ router.put("/:id", checkAuth, multer({ storage: storage }).single("image"), (req
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
   console.log(post);
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    // console.log(result);
-    res.status(200).json({ message: "Update successful!" });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: "Update successful!" });
+    } else {
+      res.status(401).json({ message: "Not authorized!" });
+    }
   });
 });
 
@@ -78,8 +83,8 @@ router.get("", (req, res, next) => {
   const currentPage = +req.query.page;
   const postQuery = Post.find();      //eseguita solo quando si invoca then()
   let fetchedPosts;
-  if(pageSize && currentPage){
-    postQuery.skip(pageSize * (currentPage-1)).limit(pageSize);
+  if (pageSize && currentPage) {
+    postQuery.skip(pageSize * (currentPage - 1)).limit(pageSize);
   }
   postQuery.then(documents => {
     fetchedPosts = documents;
@@ -116,9 +121,13 @@ router.get("/:id", (req, res, next) => {
 
 router.delete("/:id", checkAuth, (req, res, next) => {
   // console.log(req.params.id);
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    // console.log(result);
+    if (result.n > 0) {
+      res.status(200).json({ message: "Deletion successful!" });
+    } else {
+      res.status(401).json({ message: "Not authorized!" });
+    }
   });
 });
